[springframework-data-elasticsearch], [springframework-data-mongo] bulk test2

+++ springframework-data-mongo/src/main/java/hello/CustomerOp.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface CustomerOp {
 
-    int updateCompany(String name, int index, String companyName);
+    long updateCompany(String name, int index, String companyName);
 
     void findUsingName(String name);
